requirements.txt
tweepy==3.10.0
python-decouple
lyricsgenius